added: menu-graphics TU
