+++ src/headers/game-graphics.hpp
@@ -26,13 +26,6 @@ std::string BoardSizeErrorPrompt();
 std::string InputCommandListPrompt();
 std::string EndlessModeCommandListPrompt();
 std::string InputCommandListFooterPrompt();
-namespace Menu {
-std::string MainMenuTitlePrompt();
-std::string MainMenuOptionsPrompt();
-std::string InputMenuErrorInvalidInputPrompt();
-std::string InputMenuPrompt();
-std::string MainMenuGraphicsOverlay(bool input_error_choice_invalid);
-} // namespace Menu
 } // namespace Graphics
 } // namespace Game
